addressed stream comments
